+++ R/permshap.R
@@ -5,7 +5,7 @@
 #'
 #' @inheritParams kernelshap
 #' @returns
-#'   An object of class "permshap" with the following components:
+#'   An object of class "kernelshap" with the following components:
 #'   - `S`: \eqn{(n \times p)} matrix with SHAP values or, if the model output has
 #'     dimension \eqn{K > 1}, a list of \eqn{K} such matrices.
 #'   - `X`: Same as input argument `X`.
@@ -16,6 +16,7 @@
 #'     (currently `TRUE`).
 #'   - `txt`: Summary text.
 #'   - `predictions`: \eqn{(n \times K)} matrix with predictions of `X`.
+#'   - `algorithm`: "permshap".
 #' @references
 #'   1. Erik Strumbelj and Igor Kononenko. Explaining prediction models and individual 
 #'     predictions with feature contributions. Knowledge and Information Systems 41, 2014.
@@ -141,9 +142,10 @@ permshap.default <- function(object, X, bg_X, pred_fun = stats::predict,
     m_exact = m_exact,
     exact = TRUE,
     txt = txt,
-    predictions = v1
+    predictions = v1,
+    algorithm = "permshap"
   )
-  class(out) <- "permshap"
+  class(out) <- "kernelshap"
   out
 }