+++ tests/testthat/test-methods.R
@@ -15,14 +15,18 @@ test_that("print() and summary() do not give an error", {
   capture_output(expect_no_error(summary(s, compact = TRUE)))
 })
 
-# permshap
-s <- permshap(fit, iris[1:2, -1L], bg_X = iris[-1L], verbose = FALSE)
-test_that("is.permshap() works", {
-  expect_true(is.permshap(s))
-  expect_false(is.permshap(1))
+test_that("is.kernelshap(), print() and summary() also works for permshap", {
+  s <- permshap(fit, iris[1:2, -1L], bg_X = iris[-1L], verbose = FALSE)
+  
+  expect_true(is.kernelshap(s))
+  capture_output(expect_no_error(print(s)))
+  capture_output(expect_no_error(summary(s)))
 })
 
-test_that("print() and summary() also works for permshap", {
+test_that("is.kernelshap(), print() and summary() also works for additive_shap", {
+  s <- additive_shap(fit, iris, verbose = FALSE)
+  
+  expect_true(is.kernelshap(s))
   capture_output(expect_no_error(print(s)))
   capture_output(expect_no_error(summary(s)))
 })